fix(jpa): Corregir columnDefinition en entidades catalog para schema validation
PROBLEMA:
- Hibernate schema validation fallaba al iniciar la aplicación
- Mismatch entre @Column columnDefinition y tipos reales en PostgreSQL

CAMBIOS:
1. SensorType.kt (líneas 54, 62):
   - min_value: DECIMAL(10,2) → NUMERIC(10,2)
   - max_value: DECIMAL(10,2) → NUMERIC(10,2)
   - Razón: PostgreSQL reporta NUMERIC, no DECIMAL (son sinónimos pero Hibernate valida string literal)

2. AlertSeverity.kt (línea 57):
   - notification_delay_minutes: SMALLINT → INTEGER
   - Razón: Migration V12 línea 187 usa INT en lugar de SMALLINT

IMPACTO:
- ✅ Resuelve CrashLoopBackOff en pods Kubernetes
- ✅ API puede iniciar correctamente con hibernate.ddl-auto=validate
- ✅ Cero cambios en base de datos (solo ajuste de metadata JPA)

+++ src/main/kotlin/com/apptolast/invernaderos/entities/metadata/catalog/AlertSeverity.kt
@@ -50,8 +50,11 @@ data class AlertSeverity(
 
     /**
      * Minutos de delay antes de notificar
+     * IMPORTANTE: DB tiene INTEGER (4 bytes) en lugar de SMALLINT (2 bytes)
+     * Error en migration V12 línea 187 - usa INT en lugar de SMALLINT
+     * Trade-off: Acepta usar más espacio para evitar mismatch de schema validation
      */
-    @Column(name = "notification_delay_minutes", columnDefinition = "SMALLINT")
+    @Column(name = "notification_delay_minutes", columnDefinition = "INTEGER")
     val notificationDelayMinutes: Short = 0,
 
     /**
